app/requirements.txt
django==5.1.7
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
celery==5.5.2
django-celery-beat==2.8.0
redis==5.2.1
